+++ src/main/java/com/LetMeDoWith/LetMeDoWith/client/AuthClient.java
@@ -1,6 +1,6 @@
 package com.LetMeDoWith.LetMeDoWith.client;
 
-import com.LetMeDoWith.LetMeDoWith.dto.responseDto.OidcPublicKeyResDto;
+import com.LetMeDoWith.LetMeDoWith.dto.responseDto.client.OidcPublicKeyResDto;
 import reactor.core.publisher.Mono;
 
 public interface AuthClient {
