--- src/main/java/com/LetMeDoWith/LetMeDoWith/dto/responseDto/OidcPublicKeyResDto.java
@@ -1,11 +1,0 @@
-package com.LetMeDoWith.LetMeDoWith.dto.responseDto;
-
-import com.LetMeDoWith.LetMeDoWith.dto.valueObject.OidcPublicKeyVO;
-import java.util.List;
-import lombok.Builder;
-
-@Builder
-public record OidcPublicKeyResDto(List<OidcPublicKeyVO> keys) {
-    
-
-}
