--- src/main/java/com/LetMeDoWith/LetMeDoWith/dto/valueObject/OidcPublicKeyVO.java
@@ -1,10 +1,0 @@
-package com.LetMeDoWith.LetMeDoWith.dto.valueObject;
-
-public record OidcPublicKeyVO(String kid,
-                              String alg,
-                              String use,
-                              String kty,
-                              String n,
-                              String e) {
-    
-}
